refactor(package): Make prepare for publish a higher level operation
